# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CPIM-MES API",
    "description": "CPIM準拠の生産管理システムのREST API仕様。認証は JWT Bearer Token。\nログインで取得したトークンを Authorize ボタンから設定すれば、保護エンドポイントを試行できます。",
    "version": "1.0.0",
    "contact": {
      "name": "CPIM-MES"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "本サーバ"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION"
          },
          "message": {
            "type": "string",
            "example": "validation failed"
          },
          "details": {}
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "admin123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOi..."
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "planner",
              "operator",
              "viewer"
            ]
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "ログイン",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/items/import": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "CSVインポート",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  }
}